show articles on game page
