+++ src/components/AwardBlurb.astro
@@ -2,7 +2,7 @@
 import { medals } from "../airtable/data/common";
 import { type AwardDetails } from "../awards";
 import { buildAwardUrl } from "../awards";
-import type { Category } from "../utils";
+import type { Category } from "../utils/data";
 import ExternalLink from "./links/ExternalLink.astro";
 
 export type Props = AwardDetails & { category: Category };
