+++ src/components/MovieLegend.astro
@@ -1,5 +1,5 @@
 ---
-import { collectionPermalink } from "../utils";
+import { collectionPermalink } from "../utils/data";
 import ExternalLink from "./links/ExternalLink.astro";
 import RoundedLink from "./links/RoundedLink.astro";
