+++ src/components/SubscribeBlurb.astro
@@ -1,5 +1,5 @@
 ---
-import type { Category } from "../utils";
+import type { Category } from "../utils/data";
 import RssIcon from "./icons/rss.astro";
 import SubtleLink from "./links/SubtleLink.astro";
