+++ src/components/infoboxes/BaseBox.astro
@@ -1,5 +1,5 @@
 ---
-import { NO_REVIEW, type Category } from "../../utils";
+import { NO_REVIEW, type Category } from "../../utils/data";
 import Rating from "../Rating.astro";
 import BookIcon from "../icons/book.astro";
 import GameIcon from "../icons/game.astro";
