+++ src/components/infoboxes/PlayBox.astro
@@ -3,7 +3,7 @@ import { Image } from "astro:assets";
 
 import { medals } from "../../airtable/data/common";
 import type { Play } from "../../airtable/data/plays";
-import { minutesToDuration } from "../../utils";
+import { minutesToDuration } from "../../utils/data";
 import SubtleLink from "../links/SubtleLink.astro";
 import TitleLink from "../links/TitleLink.astro";
 import BaseBox from "./BaseBox.astro";
