+++ src/components/infoboxes/PlaySummary.astro
@@ -1,6 +1,6 @@
 ---
 import type { Play } from "../../airtable/data/plays";
-import { NO_REVIEW, minutesToDuration } from "../../utils";
+import { NO_REVIEW, minutesToDuration } from "../../utils/data";
 import Rating from "../Rating.astro";
 
 type Props = {
