+++ src/content/articles/beast-breaker-review.mdx
@@ -1,5 +1,5 @@
 ---
-title: "Beast Breaker is an indie billiards-esque RPG"
+title: "Beast Breaker, a Billiards Mouse RPG"
 ogDesc: "What if you were playing a billiards RPG, but your cue ball was you, an adorable warrior mouse armed with a variety of pointy objects?"
 publishedOn: "2021-09-29"
 gameSlugs:
