+++ src/layouts/Layout.astro
@@ -3,7 +3,7 @@ import { SEO, type Props as SEOProps } from "astro-seo";
 
 import Navbar from "../components/Navbar.astro";
 import type { ExternalUrl } from "../airtable/types";
-import type { Category } from "../utils";
+import type { Category } from "../utils/data";
 
 type ImageInfo = {
   url: ExternalUrl;
