+++ src/pages/articles/[slug].astro
@@ -2,7 +2,7 @@
 import ExternalLink from "../../components/links/ExternalLink.astro";
 import RoundedLink from "../../components/links/RoundedLink.astro";
 import Layout from "../../layouts/Layout.astro";
-import { seoTitle } from "../../utils";
+import { seoTitle } from "../../utils/data";
 import { getCollection, type CollectionEntry } from "astro:content";
 import CalendarIcon from "../../components/icons/calendar.astro";
