+++ src/pages/articles/index.astro
@@ -1,10 +1,10 @@
 ---
-import { getCollection } from "astro:content";
 import SubtleLink from "../../components/links/SubtleLink.astro";
 import Layout from "../../layouts/Layout.astro";
-import { seoTitle } from "../../utils";
+import { articlePermalink, seoTitle } from "../../utils/data";
+import { getPublishedArticles } from "../../utils/content";
 
-const allArticles = await getCollection("articles");
+const allArticles = await getPublishedArticles();
 ---
 
 <Layout
@@ -29,7 +29,10 @@ const allArticles = await getCollection("articles");
       allArticles.map((article) => (
         <div class="py-3">
           <div class="text-sm text-zinc-400">{article.data.publishedOn}</div>
-          <SubtleLink href={`/articles/${article.slug}`} classes={["text-xl"]}>
+          <SubtleLink
+            href={articlePermalink(article.slug)}
+            classes={["text-xl"]}
+          >
             {article.data.title}
           </SubtleLink>
         </div>
