+++ src/pages/books/[bookSlug].astro
@@ -9,7 +9,7 @@ import ExternalLink from "../../components/links/ExternalLink.astro";
 import RoundedLink from "../../components/links/RoundedLink.astro";
 import Rating from "../../components/Rating.astro";
 import Layout from "../../layouts/Layout.astro";
-import { NO_REVIEW, pluralize, seoTitle } from "../../utils";
+import { NO_REVIEW, pluralize, seoTitle } from "../../utils/data";
 
 type Params = {
   bookSlug: string;
