+++ src/pages/books/authors/[authorSlug].astro
@@ -5,7 +5,7 @@ import BookLegend from "../../../components/BookLegend.astro";
 import RoundedText from "../../../components/RoundedText.astro";
 import ReadBox from "../../../components/infoboxes/ReadBox.astro";
 import Layout from "../../../layouts/Layout.astro";
-import { averageRating, pluralize, seoTitle } from "../../../utils";
+import { averageRating, pluralize, seoTitle } from "../../../utils/data";
 
 type Params = {
   authorSlug: string;
