+++ src/pages/books/series/[seriesSlug].astro
@@ -8,7 +8,7 @@ import ExternalLink from "../../../components/links/ExternalLink.astro";
 import Rating from "../../../components/Rating.astro";
 import RoundedText from "../../../components/RoundedText.astro";
 import Layout from "../../../layouts/Layout.astro";
-import { averageRating, pluralize, seoTitle } from "../../../utils";
+import { averageRating, pluralize, seoTitle } from "../../../utils/data";
 
 type Params = {
   seriesSlug: string;
