+++ src/pages/games/[...slug].astro
@@ -6,7 +6,7 @@ import SubscribeBlurb from "../../components/SubscribeBlurb.astro";
 import PlayBox from "../../components/infoboxes/PlayBox.astro";
 import RoundedLink from "../../components/links/RoundedLink.astro";
 import Layout from "../../layouts/Layout.astro";
-import { genrePermalink, seoTitle } from "../../utils";
+import { genrePermalink, seoTitle } from "../../utils/data";
 
 type Subpaths = "awarded" | "all" | "recommended" | "recent" | "unrecommended";
 type PlayFilter = (p: Play) => boolean;
