+++ src/pages/games/collections/[collectionSlug].astro
@@ -15,7 +15,7 @@ import {
   averageRating,
   collectionPermalink,
   seoTitle,
-} from "../../../utils";
+} from "../../../utils/data";
 
 type Params = {
   collectionSlug: string;
