+++ src/pages/games/genres/[genreSlug].astro
@@ -11,7 +11,7 @@ import RoundedText from "../../../components/RoundedText.astro";
 import RoundedLink from "../../../components/links/RoundedLink.astro";
 import TitleLink from "../../../components/links/TitleLink.astro";
 import Layout from "../../../layouts/Layout.astro";
-import { genrePermalink, seoTitle } from "../../../utils";
+import { genrePermalink, seoTitle } from "../../../utils/data";
 
 type Params = {
   genreSlug: string;
