+++ src/pages/games/years/[year].astro
@@ -7,7 +7,7 @@ import PlayBox from "../../../components/infoboxes/PlayBox.astro";
 import ExternalLink from "../../../components/links/ExternalLink.astro";
 import RoundedLink from "../../../components/links/RoundedLink.astro";
 import Layout from "../../../layouts/Layout.astro";
-import { genrePermalink, seoTitle } from "../../../utils";
+import { genrePermalink, seoTitle } from "../../../utils/data";
 
 type Params = { year: string };
