+++ src/pages/movies/collections/[collectionSlug].astro
@@ -12,7 +12,7 @@ import {
   collectionPermalink,
   pluralize,
   seoTitle,
-} from "../../../utils";
+} from "../../../utils/data";
 
 type Params = {
   collectionSlug: string;
