+++ src/pages/movies/years/[year].astro
@@ -6,7 +6,7 @@ import SubscribeBlurb from "../../../components/SubscribeBlurb.astro";
 import WatchBox from "../../../components/infoboxes/WatchBox.astro";
 import ExternalLink from "../../../components/links/ExternalLink.astro";
 import Layout from "../../../layouts/Layout.astro";
-import { seoTitle } from "../../../utils";
+import { seoTitle } from "../../../utils/data";
 
 type Params = { year: string };
