+++ src/pages/search.astro
@@ -8,7 +8,7 @@ import { loadMovies } from "../airtable/data/movies";
 import { loadSeries } from "../airtable/data/series";
 import { SearchBar } from "../components/SearchBar";
 import Layout from "../layouts/Layout.astro";
-import { type SearchItem } from "../utils";
+import { type SearchItem } from "../utils/search";
 
 const list: SearchItem[] = [
   // movies
