+++ src/utils/search.ts
@@ -1,0 +1,7 @@
+import { type Category } from "./data";
+
+export type SearchItem = {
+  title: string;
+  category: Category;
+  permalink: string;
+};
